feat: add pre-commit framework with clang-format setup
The pre-commit framework provides a CLI tool to set up a collection of
hooks that execute on changed files, or optionally on all files.

This setup includes a hook to run clang-format, ensuring that
LeanStore's style conventions are easy to adhere to. For initial setup,
it uses the Google style.

Additionally, a helper command 'make format' is added to simplify
formatting.

Signed-off-by: Simon A. F. Lund <os@safl.dk>